+++ src/components/Table/Table.tsx
@@ -16,14 +16,14 @@ type TableCompoundProps = {
 };
 
 const Table: FC<Props> & TableCompoundProps = (props) => {
-  const { columns, rows, emptyState, sorting, onRowSelect } = props;
+  const { columns, rows, emptyState, onRowSelect, sorting } = props;
 
   const [state, dispatch] = useReducer(reducer, {
     columns,
     rows,
     emptyState,
-    selected: [],
     sorting,
+    selected: [],
   });
 
   const { selected } = state;